ref(BuilderDrawerContext): move clear active field logic into context
