+++ frontend/src/features/admin-form-builder/BuilderDrawerContext.tsx
@@ -8,7 +8,11 @@ import {
   useState,
 } from 'react'
 
-import { activeFieldSelector, useEditFieldStore } from './editFieldStore'
+import {
+  activeFieldSelector,
+  clearActiveFieldSelector,
+  useEditFieldStore,
+} from './editFieldStore'
 
 export enum DrawerTabs {
   Builder,
@@ -59,6 +63,10 @@ export const useBuilderDrawer = (): BuilderDrawerContextProps => {
 const useProvideDrawerContext = (): BuilderDrawerContextProps => {
   const [activeTab, setActiveTab] = useState<DrawerTabs | null>(null)
   const activeField = useEditFieldStore(activeFieldSelector)
+  const hasActiveField = useEditFieldStore(
+    useCallback((state) => !!state.activeField, []),
+  )
+  const clearActiveField = useEditFieldStore(clearActiveFieldSelector)
 
   useEffect(() => {
     if (activeField) {
@@ -71,7 +79,12 @@ const useProvideDrawerContext = (): BuilderDrawerContextProps => {
     [activeTab],
   )
 
-  const handleClose = useCallback(() => setActiveTab(null), [])
+  const handleClose = useCallback(() => {
+    if (hasActiveField) {
+      clearActiveField()
+    }
+    setActiveTab(null)
+  }, [clearActiveField, hasActiveField])
 
   const handleBuilderClick = () => setActiveTab(DrawerTabs.Builder)
   const handleDesignClick = () => setActiveTab(DrawerTabs.Design)
