+++ frontend/src/features/admin-form-builder/FieldRow/BuilderDrawerCloseButton.tsx
@@ -1,23 +1,10 @@
-import { useCallback } from 'react'
 import { BiX } from 'react-icons/bi'
 import { CloseButton } from '@chakra-ui/react'
 
 import { useBuilderDrawer } from '../BuilderDrawerContext'
-import { clearActiveFieldSelector, useEditFieldStore } from '../editFieldStore'
 
 export const BuilderDrawerCloseButton = (): JSX.Element => {
   const { handleClose } = useBuilderDrawer()
-  const hasActiveField = useEditFieldStore(
-    useCallback((state) => !!state.activeField, []),
-  )
-  const clearActiveField = useEditFieldStore(clearActiveFieldSelector)
-
-  const handleDrawerClose = useCallback(() => {
-    if (hasActiveField) {
-      clearActiveField()
-    }
-    handleClose()
-  }, [clearActiveField, handleClose, hasActiveField])
 
   return (
     <CloseButton
@@ -28,7 +15,7 @@ export const BuilderDrawerCloseButton = (): JSX.Element => {
       variant="clear"
       colorScheme="neutral"
       children={<BiX />}
-      onClick={handleDrawerClose}
+      onClick={handleClose}
     />
   )
 }
